+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 All notable changes to this project will be documented in this file.
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## [2.2.1] --- in progress
+
+## [2.3.0] --- in progress
 ### Added
 - Build Info Details editorscript that shows the build info in the editor
 - PackageName to build info
+- [ReadMeHelpers v2.0.0](https://github.com/solo-fsw/sosxr-unity-readmehelpers)
+
 
 ### Fixed
 - Directory structure from the FileHelpers